// frontend/src/generated/models/GenerateJsonWebKeyProperties.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Small KMS Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { JsonWebKeyCurveName } from './JsonWebKeyCurveName';
import {
    JsonWebKeyCurveNameFromJSON,
    JsonWebKeyCurveNameFromJSONTyped,
    JsonWebKeyCurveNameToJSON,
} from './JsonWebKeyCurveName';
import type { JsonWebKeyOperation } from './JsonWebKeyOperation';
import {
    JsonWebKeyOperationFromJSON,
    JsonWebKeyOperationFromJSONTyped,
    JsonWebKeyOperationToJSON,
} from './JsonWebKeyOperation';
import type { JsonWebKeyType } from './JsonWebKeyType';
import {
    JsonWebKeyTypeFromJSON,
    JsonWebKeyTypeFromJSONTyped,
    JsonWebKeyTypeToJSON,
} from './JsonWebKeyType';

/**
 * 
 * @export
 * @interface GenerateJsonWebKeyProperties
 */
export interface GenerateJsonWebKeyProperties {
    /**
     * 
     * @type {JsonWebKeyType}
     * @memberof GenerateJsonWebKeyProperties
     */
    kty?: JsonWebKeyType;
    /**
     * 
     * @type {JsonWebKeyCurveName}
     * @memberof GenerateJsonWebKeyProperties
     */
    crv?: JsonWebKeyCurveName;
    /**
     * 
     * @type {number}
     * @memberof GenerateJsonWebKeyProperties
     */
    keySize?: number;
    /**
     * 
     * @type {Array<JsonWebKeyOperation>}
     * @memberof GenerateJsonWebKeyProperties
     */
    keyOps?: Array<JsonWebKeyOperation>;
}

/**
 * Check if a given object implements the GenerateJsonWebKeyProperties interface.
 */
export function instanceOfGenerateJsonWebKeyProperties(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GenerateJsonWebKeyPropertiesFromJSON(json: any): GenerateJsonWebKeyProperties {
    return GenerateJsonWebKeyPropertiesFromJSONTyped(json, false);
}

export function GenerateJsonWebKeyPropertiesFromJSONTyped(json: any, ignoreDiscriminator: boolean): GenerateJsonWebKeyProperties {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'kty': !exists(json, 'kty') ? undefined : JsonWebKeyTypeFromJSON(json['kty']),
        'crv': !exists(json, 'crv') ? undefined : JsonWebKeyCurveNameFromJSON(json['crv']),
        'keySize': !exists(json, 'key_size') ? undefined : json['key_size'],
        'keyOps': !exists(json, 'key_ops') ? undefined : ((json['key_ops'] as Array<any>).map(JsonWebKeyOperationFromJSON)),
    };
}

export function GenerateJsonWebKeyPropertiesToJSON(value?: GenerateJsonWebKeyProperties | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'kty': JsonWebKeyTypeToJSON(value.kty),
        'crv': JsonWebKeyCurveNameToJSON(value.crv),
        'key_size': value.keySize,
        'key_ops': value.keyOps === undefined ? undefined : ((value.keyOps as Array<any>).map(JsonWebKeyOperationToJSON)),
    };
}