// frontend/src/generated/models/KeyPolicy.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Small KMS Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { GenerateJsonWebKeyProperties } from './GenerateJsonWebKeyProperties';
import {
    GenerateJsonWebKeyPropertiesFromJSON,
    GenerateJsonWebKeyPropertiesFromJSONTyped,
    GenerateJsonWebKeyPropertiesToJSON,
} from './GenerateJsonWebKeyProperties';

/**
 * 
 * @export
 * @interface KeyPolicy
 */
export interface KeyPolicy {
    /**
     * 
     * @type {string}
     * @memberof KeyPolicy
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof KeyPolicy
     */
    updated: Date;
    /**
     * 
     * @type {Date}
     * @memberof KeyPolicy
     */
    deleted?: Date;
    /**
     * 
     * @type {string}
     * @memberof KeyPolicy
     */
    updatedBy?: string;
    /**
     * 
     * @type {string}
     * @memberof KeyPolicy
     */
    displayName: string;
    /**
     * 
     * @type {GenerateJsonWebKeyProperties}
     * @memberof KeyPolicy
     */
    keyProperties: GenerateJsonWebKeyProperties;
    /**
     * 
     * @type {boolean}
     * @memberof KeyPolicy
     */
    exportable: boolean;
    /**
     * 
     * @type {string}
     * @memberof KeyPolicy
     */
    expiryTime?: string;
}

/**
 * Check if a given object implements the KeyPolicy interface.
 */
export function instanceOfKeyPolicy(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "updated" in value;
    isInstance = isInstance && "displayName" in value;
    isInstance = isInstance && "keyProperties" in value;
    isInstance = isInstance && "exportable" in value;

    return isInstance;
}

export function KeyPolicyFromJSON(json: any): KeyPolicy {
    return KeyPolicyFromJSONTyped(json, false);
}

export function KeyPolicyFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeyPolicy {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'updated': (new Date(json['updated'])),
        'deleted': !exists(json, 'deleted') ? undefined : (new Date(json['deleted'])),
        'updatedBy': !exists(json, 'updatedBy') ? undefined : json['updatedBy'],
        'displayName': json['displayName'],
        'keyProperties': GenerateJsonWebKeyPropertiesFromJSON(json['keyProperties']),
        'exportable': json['exportable'],
        'expiryTime': !exists(json, 'expiryTime') ? undefined : json['expiryTime'],
    };
}

export function KeyPolicyToJSON(value?: KeyPolicy | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'updated': (value.updated.toISOString()),
        'deleted': value.deleted === undefined ? undefined : (value.deleted.toISOString()),
        'updatedBy': value.updatedBy,
        'displayName': value.displayName,
        'keyProperties': GenerateJsonWebKeyPropertiesToJSON(value.keyProperties),
        'exportable': value.exportable,
        'expiryTime': value.expiryTime,
    };
}