// frontend/src/generated/models/KeyPolicyParameters.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Small KMS Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { GenerateJsonWebKeyProperties } from './GenerateJsonWebKeyProperties';
import {
    GenerateJsonWebKeyPropertiesFromJSON,
    GenerateJsonWebKeyPropertiesFromJSONTyped,
    GenerateJsonWebKeyPropertiesToJSON,
} from './GenerateJsonWebKeyProperties';

/**
 * 
 * @export
 * @interface KeyPolicyParameters
 */
export interface KeyPolicyParameters {
    /**
     * 
     * @type {string}
     * @memberof KeyPolicyParameters
     */
    displayName?: string;
    /**
     * 
     * @type {GenerateJsonWebKeyProperties}
     * @memberof KeyPolicyParameters
     */
    keyProperties?: GenerateJsonWebKeyProperties;
    /**
     * 
     * @type {boolean}
     * @memberof KeyPolicyParameters
     */
    exportable?: boolean;
    /**
     * 
     * @type {string}
     * @memberof KeyPolicyParameters
     */
    expiryTime?: string;
}

/**
 * Check if a given object implements the KeyPolicyParameters interface.
 */
export function instanceOfKeyPolicyParameters(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function KeyPolicyParametersFromJSON(json: any): KeyPolicyParameters {
    return KeyPolicyParametersFromJSONTyped(json, false);
}

export function KeyPolicyParametersFromJSONTyped(json: any, ignoreDiscriminator: boolean): KeyPolicyParameters {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'displayName': !exists(json, 'displayName') ? undefined : json['displayName'],
        'keyProperties': !exists(json, 'keyProperties') ? undefined : GenerateJsonWebKeyPropertiesFromJSON(json['keyProperties']),
        'exportable': !exists(json, 'exportable') ? undefined : json['exportable'],
        'expiryTime': !exists(json, 'expiryTime') ? undefined : json['expiryTime'],
    };
}

export function KeyPolicyParametersToJSON(value?: KeyPolicyParameters | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'displayName': value.displayName,
        'keyProperties': GenerateJsonWebKeyPropertiesToJSON(value.keyProperties),
        'exportable': value.exportable,
        'expiryTime': value.expiryTime,
    };
}